transform: move Propagate impl in components/mod.rs
